+++ src/main/java/com/refinedmods/refinedsites/model/release/Platform.java
@@ -1,0 +1,7 @@
+package com.refinedmods.refinedsites.model.release;
+
+public enum Platform {
+    NEOFORGE,
+    FORGE,
+    FABRIC
+}